refactor: stop exporting NativeScriptModule from platform
Importing NativeScriptModule from "nativescript-angular/platform" leads to
including @angular/compiler in the bundle when doing AoT compilation
with webpack. Import from "nativescript-angular/nativescript.module"
should be used instead.

BREAKING CHANGE: User applications cannot import NativeScriptModule from
"nativescript-angular/platform" anymore.
Migration:
Before:
```
import { NativeScriptModule } from "nativescript-angular/platform";
```
After
```
import { NativeScriptModule } from
"nativescript-angular/nativescript.module";
```
