+++ tests/app/lazy-loaded.module.ts
@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import { NativeScriptModule } from "nativescript-angular";
+import { NativeScriptModule } from "nativescript-angular/nativescript.module";
 import { NativeScriptRouterModule } from "nativescript-angular/router";
 
 import { SecondComponent } from './second.component';
