+++ tests/app/snippets/navigation/page-outlet.ts
@@ -2,7 +2,8 @@ import { TestApp, registerTestApp } from "../../tests/test-app";
 import { ApplicationRef } from '@angular/core';
 import { Router, NavigationStart, NavigationEnd } from "@angular/router";
 // >> page-outlet-example
-import { platformNativeScriptDynamic, NativeScriptModule } from "nativescript-angular/platform";
+import { platformNativeScriptDynamic } from "nativescript-angular/platform";
+import { NativeScriptModule } from "nativescript-angular/nativescript.module";
 import { Component, NgModule } from '@angular/core';
 import { NativeScriptRouterModule } from "nativescript-angular/router";
 import { routes } from "./app.routes";
